scraper/requirements.txt
atproto
minio
chromadb
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers
apscheduler
requests
python-dotenv
structlog
